+++ packages/yourloops/components/monitoring-alert/monitoring-alerts-card.tsx
@@ -41,6 +41,7 @@ import Typography from '@mui/material/Typography'
 import { useAuth } from '../../lib/auth'
 import { useTheme } from '@mui/material/styles'
 import { MonitoringAlertsCardSkeletonValue } from './monitoring-alerts-card-skeleton-value'
+import PatientUtils from '../../lib/patient/patient.util'
 
 const monitoringAlertsCardStyles = makeStyles()((theme) => {
   return {
@@ -117,7 +118,7 @@ function MonitoringAlertsCard(props: MonitoringAlertsCardProps): JSX.Element {
             >
               {t('time-out-of-range-target')}
               <Box>
-                {`${Math.round(patient.monitoringAlerts.timeSpentAwayFromTargetRate * 10) / 10}%`}
+                {PatientUtils.formatPercentageValue(patient.monitoringAlerts.timeSpentAwayFromTargetRate)}
               </Box>
             </Box>
             <Box
@@ -129,7 +130,7 @@ function MonitoringAlertsCard(props: MonitoringAlertsCardProps): JSX.Element {
             >
               {t('alert-hypoglycemic')}
               <Box>
-                {`${Math.round(patient.monitoringAlerts.frequencyOfSevereHypoglycemiaRate * 10) / 10}%`}
+                {PatientUtils.formatPercentageValue(patient.monitoringAlerts.frequencyOfSevereHypoglycemiaRate)}
               </Box>
             </Box>
             <Box
@@ -140,7 +141,9 @@ function MonitoringAlertsCard(props: MonitoringAlertsCardProps): JSX.Element {
               className={nonDataTransmissionActive ? classes.alertColor : ''}
             >
               {t('data-not-transmitted')}
-              <Box>{`${Math.round(patient.monitoringAlerts.nonDataTransmissionRate * 10) / 10}%`}</Box>
+              <Box>
+                {PatientUtils.formatPercentageValue(patient.monitoringAlerts.nonDataTransmissionRate)}
+              </Box>
             </Box>
           </>
           : <Box>